backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pandas
flask

dnspython
httpx

